TVIST1-297: Replaced relation from BoardMember to Municipality with BoardMember to Board
