+++ src/Controller/Admin/BoardMemberCrudController.php
@@ -29,7 +29,7 @@ class BoardMemberCrudController extends AbstractCrudController
     public function configureFields(string $pageName): iterable
     {
         yield TextField::new('name', 'Name');
-        yield AssociationField::new('municipality', 'Municipality')
+        yield AssociationField::new('boards', 'Board')
             ->setRequired(true);
         yield AssociationField::new('boardRoles', 'BoardRole')
             ->setFormTypeOptions([
