+++ src/Entity/Board.php
@@ -72,12 +72,18 @@ class Board implements LoggableEntityInterface
      */
     private $agendas;
 
+    /**
+     * @ORM\ManyToMany(targetEntity=BoardMember::class, mappedBy="boards")
+     */
+    private $boardMembers;
+
     public function __construct()
     {
         $this->complaintCategories = new ArrayCollection();
         $this->caseEntities = new ArrayCollection();
         $this->boardRoles = new ArrayCollection();
         $this->agendas = new ArrayCollection();
+        $this->boardMembers = new ArrayCollection();
     }
 
     public function getId(): ?UuidV4
@@ -275,6 +281,33 @@ class Board implements LoggableEntityInterface
             if ($agenda->getBoard() === $this) {
                 $agenda->setBoard(null);
             }
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|BoardMember[]
+     */
+    public function getBoardMembers(): Collection
+    {
+        return $this->boardMembers;
+    }
+
+    public function addBoardMember(BoardMember $boardMember): self
+    {
+        if (!$this->boardMembers->contains($boardMember)) {
+            $this->boardMembers[] = $boardMember;
+            $boardMember->addBoard($this);
+        }
+
+        return $this;
+    }
+
+    public function removeBoardMember(BoardMember $boardMember): self
+    {
+        if ($this->boardMembers->removeElement($boardMember)) {
+            $boardMember->removeBoard($this);
         }
 
         return $this;
