+++ src/Entity/BoardMember.php
@@ -28,12 +28,6 @@ class BoardMember
     private $name;
 
     /**
-     * @ORM\ManyToOne(targetEntity=Municipality::class, inversedBy="boardMembers")
-     * @ORM\JoinColumn(nullable=false)
-     */
-    private $municipality;
-
-    /**
      * @ORM\ManyToMany(targetEntity=BoardRole::class, mappedBy="boardMembers")
      */
     private $boardRoles;
@@ -43,10 +37,16 @@ class BoardMember
      */
     private $agendas;
 
+    /**
+     * @ORM\ManyToMany(targetEntity=Board::class, inversedBy="boardMembers")
+     */
+    private $boards;
+
     public function __construct()
     {
         $this->boardRoles = new ArrayCollection();
         $this->agendas = new ArrayCollection();
+        $this->boards = new ArrayCollection();
     }
 
     public function getId(): ?UuidV4
@@ -62,18 +62,6 @@ class BoardMember
     public function setName(string $name): self
     {
         $this->name = $name;
-
-        return $this;
-    }
-
-    public function getMunicipality(): ?Municipality
-    {
-        return $this->municipality;
-    }
-
-    public function setMunicipality(?Municipality $municipality): self
-    {
-        $this->municipality = $municipality;
 
         return $this;
     }
@@ -135,5 +123,29 @@ class BoardMember
     public function __toString()
     {
         return $this->name;
+    }
+
+    /**
+     * @return Collection|Board[]
+     */
+    public function getBoards(): Collection
+    {
+        return $this->boards;
+    }
+
+    public function addBoard(Board $board): self
+    {
+        if (!$this->boards->contains($board)) {
+            $this->boards[] = $board;
+        }
+
+        return $this;
+    }
+
+    public function removeBoard(Board $board): self
+    {
+        $this->boards->removeElement($board);
+
+        return $this;
     }
 }
